Ready for Review
Ready for Review: Added author and date, renamed what seemed to be the most home file to index.html and changed links to go to it. imported designs from the design style sheet. Then entire body is now a flex box, and many elements are also flexboxes. Most flex box related changes are to the flex grow, flex shrink and basis values. While on mobile the nav is at the bottom, but footer still below it. The navigation disappears while on mobile, unless the navicon is hovered over. The aside will not appear. While on desktop, the navicon will not appear. The horizontal navigation is better spaced and flexable while on desktop/tablet.